+++ unicorn2/src/commonMain/kotlin/ru/dimsuz/unicorn2/Machine.kt
@@ -5,19 +5,7 @@ package ru.dimsuz.unicorn2
 
 import kotlinx.coroutines.flow.Flow
 
-public interface Machine<S : Any, E : Any> {
+public interface Machine<S : Any> {
   public val initial: Pair<suspend () -> S, (suspend (S) -> Unit)?>
   public val states: Flow<S>
-
-  /**
-   * Sends event [e] to the machine, suspending in case event buffer is overflown.
-   * See [trySend] for a non-suspending variant
-   */
-  public suspend fun send(e: E)
-
-  /**
-   * Tries to send an event [e] to the machine and returns `false` if calling [send]
-   * instead of [trySend] would suspend
-   */
-  public fun trySend(e: E): Boolean
 }