+++ unicorn2/src/commonMain/kotlin/ru/dimsuz/unicorn2/MachineDsl.kt
@@ -10,59 +10,54 @@ import kotlin.reflect.KClass
 public annotation class StateMachineDsl
 
 @StateMachineDsl
-public class MachineDsl<S : Any, E : Any> @PublishedApi internal constructor(
-  public val events: Flow<E>,
+public class MachineDsl<S : Any> @PublishedApi internal constructor(
   @PublishedApi internal val klass: KClass<out S>
 ) {
   public var initial: Pair<S, (suspend (S) -> Unit)?>? = null
   public var initialLazy: Pair<suspend () -> S, (suspend (S) -> Unit)?>? = null
 
   @PublishedApi
-  internal val transitions: MutableMap<KClass<out S>, MutableList<TransitionDsl<out S, S, Any, E>>> = mutableMapOf()
+  internal val transitions: MutableMap<KClass<out S>, MutableList<TransitionDsl<out S, S, Any>>> = mutableMapOf()
 
-  public inline fun <P> onEach(eventPayloads: Flow<P>, init: TransitionDsl<S, S, P, E>.() -> Unit) {
-    val transitionDsl = TransitionDsl<S, S, P, E>(eventPayloads).apply(init)
+  public inline fun <P> onEach(eventPayloads: Flow<P>, init: TransitionDsl<S, S, P>.() -> Unit) {
+    val transitionDsl = TransitionDsl<S, S, P>(eventPayloads).apply(init)
     val list = transitions.getOrPut(klass) { arrayListOf() }
     @Suppress("UNCHECKED_CAST") // we know the types here, enforced by dsl
-    list += transitionDsl as TransitionDsl<S, S, Any, E>
+    list += transitionDsl as TransitionDsl<S, S, Any>
   }
 
-  public inline fun <reified T : S> whenIn(init: StateDsl<T, S, E>.() -> Unit) {
-    val stateDsl = StateDsl<T, S, E>(events).apply(init)
+  public inline fun <reified T : S> whenIn(init: StateDsl<T, S>.() -> Unit) {
+    val stateDsl = StateDsl<T, S>().apply(init)
     val list = transitions.getOrPut(T::class) { arrayListOf() }
     list += stateDsl.transitions
   }
 }
 
 @StateMachineDsl
-public class StateDsl<S : PS, PS : Any, E : Any>(public val events: Flow<E>) {
+public class StateDsl<S : PS, PS : Any> @PublishedApi internal constructor() {
   @PublishedApi
-  internal val transitions: MutableList<TransitionDsl<out PS, PS, Any, E>> = arrayListOf()
+  internal val transitions: MutableList<TransitionDsl<out PS, PS, Any>> = arrayListOf()
 
-  public inline fun <P> onEach(eventPayloads: Flow<P>, init: TransitionDsl<S, PS, P, E>.() -> Unit) {
-    val transitionDsl = TransitionDsl<S, PS, P, E>(eventPayloads).apply(init)
+  public inline fun <P> onEach(eventPayloads: Flow<P>, init: TransitionDsl<S, PS, P>.() -> Unit) {
+    val transitionDsl = TransitionDsl<S, PS, P>(eventPayloads).apply(init)
     @Suppress("UNCHECKED_CAST") // we know the types here, enforced by dsl
-    transitions.add(transitionDsl as TransitionDsl<S, PS, Any, E>)
+    transitions.add(transitionDsl as TransitionDsl<S, PS, Any>)
   }
 }
 
 @StateMachineDsl
-public class TransitionDsl<S : PS, PS : Any, P, E : Any> @PublishedApi internal constructor(
+public class TransitionDsl<S : PS, PS : Any, P> @PublishedApi internal constructor(
   @PublishedApi internal val eventPayloads: Flow<P>,
 ) {
 
   internal var transition: (suspend (S, P) -> PS)? = null
-  internal var action: (suspend (scope: ActionScope<E>, S, PS, P) -> Unit)? = null
+  internal var action: (suspend (S, PS, P) -> Unit)? = null
 
   public fun transitionTo(body: suspend (state: S, payload: P) -> PS) {
     this.transition = body
   }
 
-  public fun action(body: suspend ActionScope<E>.(state: S, newState: PS, payload: P) -> Unit) {
+  public fun action(body: suspend (state: S, newState: PS, payload: P) -> Unit) {
     this.action = body
   }
-}
-
-public interface ActionScope<E : Any> {
-  public fun sendEvent(event: E)
 }